pyproject.toml
[project]
name = "emu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "pymupdf>=1.26.3",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
